+++ dst/redis.js
@@ -6,7 +6,7 @@ exports.getClient = void 0;
  */
 const createDebug = require("debug");
 const redis = require("redis");
-const debug = createDebug('chevre-api:redis');
+const debug = createDebug('pecorino-api:redis');
 const CONNECT_TIMEOUT_IN_MILLISECONDS = 3600000;
 const MAX_ATTEMPTS = 10;
 const PING_INTERVAL = 60000; // 60 seconds
@@ -37,7 +37,7 @@ function createClient() {
             debug('retrying...', options);
             if (options.error instanceof Error && options.error.code === 'ECONNREFUSED') {
                 // tslint:disable-next-line:no-console
-                console.error(options.error);
+                console.error('redisClient.retry_strategy:', options.error);
                 // redisClient = redisClient.duplicate();
                 // End reconnecting on a specific error and flush all commands with a individual error
                 // return new Error('The server refused the connection');
@@ -67,7 +67,7 @@ function createClient() {
     // });
     c.on('error', (err) => {
         // tslint:disable-next-line:no-console
-        console.error(err);
+        console.error('redisClient emitted an error', err);
     });
     // c.on('end', () => {
     //     debug('end');