+++ dst/server.js
@@ -73,7 +73,7 @@ function onListening() {
     const addr = server.address();
     const bind = typeof addr === 'string'
         ? `pipe ${addr}`
-        : `port ${addr.port.toString()}`;
+        : `port ${addr === null || addr === void 0 ? void 0 : addr.port.toString()}`;
     debug(`Listening on ${bind}`);
     const diff = process.hrtime(startTime);
     debug(`api server listening took ${diff[0]} seconds and ${diff[1]} nanoseconds.`);